// projects/ngx-simple-text-editor/src/lib/models/exec-command.ts
// https://developer.mozilla.org/en-US/docs/Web/API/Document/execCommand
export const enum ExecCommand {
  backColor = 'backColor',
  bold = 'bold',
  ClearAuthenticationCache = 'ClearAuthenticationCache',
  contentReadOnly = 'contentReadOnly',
  copy = 'copy',
  createLink = 'createLink',
  cut = 'cut',
  decreaseFontSize = 'decreaseFontSize',
  defaultParagraphSeparator = 'defaultParagraphSeparator',
  delete = 'delete',
  enableAbsolutePositionEditor = 'enableAbsolutePositionEditor',
  enableInlineTableEditing = 'enableInlineTableEditing',
  enableObjectResizing = 'enableObjectResizing',
  fontName = 'fontName',
  fontSize = 'fontSize',
  foreColor = 'foreColor',
  formatBlock = 'formatBlock',
  forwardDelete = 'forwardDelete',
  heading = 'heading',
  hiliteColor = 'hiliteColor',
  increaseFontSize = 'increaseFontSize',
  indent = 'indent',
  insertBrOnReturn = 'insertBrOnReturn',
  insertHorizontalRule = 'insertHorizontalRule',
  insertHTML = 'insertHTML',
  insertImage = 'insertImage',
  insertOrderedList = 'insertOrderedList',
  insertUnorderedList = 'insertUnorderedList',
  insertParagraph = 'insertParagraph',
  insertText = 'insertText',
  italic = 'italic',
  justifyCenter = 'justifyCenter',
  justifyFull = 'justifyFull',
  justifyLeft = 'justifyLeft',
  justifyRight = 'justifyRight',
  outdent = 'outdent',
  paste = 'paste',
  redo = 'redo',
  removeFormat = 'removeFormat',
  selectAll = 'selectAll',
  strikeThrough = 'strikeThrough',
  subscript = 'subscript',
  superscript = 'superscript',
  underline = 'underline',
  undo = 'undo',
  unlink = 'unlink',
  useCSS = 'useCSS',
  styleWithCSS = 'styleWithCSS',
  AutoUrlDetect = 'AutoUrlDetect',
}